refactor: move Google Maps link generation to service class
